fix: disable nvim nvim-lint
This motherfucker was making saving files a pain in the ass, as it was
making the saving process extremely slow, I'll figure my own way of
showing these linter thingies without having to compromise on speed
every time I save a file.

NOTE: I'll probably just remove the lint on save feature, or use the
built in loclist with vimgrep, I'll figure it out.